pyproject.toml
[project]
name = "geoai"
version = "0.1.0"
description = "This is a Python package for GeoAI."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
    "langchain-chroma>=0.2.3",
    "langchain-community>=0.3.23",
    "langchain-deepseek>=0.1.3",
    "langchain-openai>=0.3.14",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.2",
]
